requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
aiohttp>=3.9.1
apscheduler>=3.10.4
python-multipart>=0.0.6
asyncpg>=0.29.0 